+++ term.go
@@ -6,11 +6,11 @@ import (
 
 type Term struct {
 	Variable    *Variable
-	Coefficient Float
+	Coefficient float64
 }
 type Terms []*Term
 
-func NewTerm(v *Variable, coefficient Float) *Term {
+func NewTerm(v *Variable, coefficient float64) *Term {
 	return &Term{
 		Variable:    v,
 		Coefficient: coefficient,
@@ -21,7 +21,7 @@ func NewTermFrom(variable *Variable) *Term {
 	return NewTerm(variable, 1.0)
 }
 
-func (t *Term) Value() Float {
+func (t *Term) Value() float64 {
 	return t.Coefficient * t.Value()
 }
 
@@ -32,11 +32,11 @@ func (t *Term) String() string {
 	)
 }
 
-func (t *Term) Multiply(coefficient Float) *Term {
+func (t *Term) Multiply(coefficient float64) *Term {
 	return NewTerm(t.Variable, t.Coefficient*coefficient)
 }
 
-func (t *Term) Divide(denominator Float) *Term {
+func (t *Term) Divide(denominator float64) *Term {
 	return t.Multiply(1 / denominator)
 }
 
@@ -57,7 +57,7 @@ func (t *Term) AddVariable(v *Variable) *Expression {
 	return t.Add(NewTermFrom(v))
 }
 
-func (t *Term) AddFloat(constant Float) *Expression {
+func (t *Term) AddFloat(constant float64) *Expression {
 	return NewExpression(constant, t)
 }
 
@@ -73,7 +73,7 @@ func (t *Term) subtract(v *Variable) *Expression {
 	return t.Add(v.Negate())
 }
 
-func (t *Term) SubtractFloat(constant Float) *Expression {
+func (t *Term) SubtractFloat(constant float64) *Expression {
 	return t.AddFloat(-constant)
 }
 
@@ -91,7 +91,7 @@ func (t *Term) EqualsVariable(v *Variable) *Constraint {
 	return NewExpressionFrom(t).EqualsVariable(v)
 }
 
-func (t *Term) EqualsFloat(constant Float) *Constraint {
+func (t *Term) EqualsFloat(constant float64) *Constraint {
 	e := NewExpressionFrom(t)
 	c := e.EqualsFloat(constant)
 	return c
@@ -115,7 +115,7 @@ func (t *Term) LessThanOrEqualToVariable(v *Variable) *Constraint {
 	return c
 }
 
-func (t *Term) LessThanOrEqualToFloat(constant Float) *Constraint {
+func (t *Term) LessThanOrEqualToFloat(constant float64) *Constraint {
 	e := NewExpressionFrom(t)
 	c := e.LessThanOrEqualToFloat(constant)
 	return c
@@ -139,7 +139,7 @@ func (t *Term) GreaterThanOrEqualToVariable(v *Variable) *Constraint {
 	return c
 }
 
-func (t *Term) GreaterThanOrEqualToFloat(constant Float) *Constraint {
+func (t *Term) GreaterThanOrEqualToFloat(constant float64) *Constraint {
 	e := NewExpressionFrom(t)
 	c := e.GreaterThanOrEqualToFloat(constant)
 	return c
